rest_api/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-environ==0.11.2
mysqlclient==2.2.4
pytest==8.3.3
pytest-django==4.9.0